address_synchronizer: add a cache in front of get_utxos()
get_utxos() is called pretty often, both spuriously,
and on focus change, on tab switch, &c.

It blocks as it iterates, functionally, /every/ address the wallet knows of.

On large wallets (like testnet
  vpub5VfkVzoT7qgd5gUKjxgGE2oMJU4zKSktusfLx2NaQCTfSeeSY3S723qXKUZZaJzaF6YaF8nwQgbMTWx54Ugkf4NZvSxdzicENHoLJh96EKg
from #6625 with 11k TXes and 10.5k addresses),
this takes 1.3s of 100%ed CPU usage,
basically in a loop from the UI thread.

get_utxos() is 50-70% of the flame-graph when sampling
a synced wallet process.

This data is a function of the block-chain state,
and we have hooks that notify us of when the block-chain state changes:
we can just cache the result and only re-compute it then.

For example, here's a trace log where get_utxos() has
  print(end - start, len(domain), block_height)
and a transaction is clearing:
  1.3775344607420266 10540 4507192
  0.0010390589013695717 10540 4507192 cached!
  0.001393263228237629 10540 4507192 cached!
  0.0009001069702208042 10540 4507192 cached!
  0.0010241391137242317 10540 4507192 cached!
  ...
  0.00207632128149271 10540 4507192 cached!
  0.001397700048983097 10540 4507192 cached!
  invalidate_cache
  1.4686454269103706 10540 4507192
  0.0012429207563400269 10540 4507192 cached!
  0.0015075239352881908 10540 4507192 cached!
  0.0010459059849381447 10540 4507192 cached!
  0.0009669591672718525 10540 4507192 cached!
  ...
  on_event_blockchain_updated
  invalidate_cache
  1.3897203942760825 10540 4507193
  0.0010689008049666882 10540 4507193 cached!
  0.0010420521721243858 10540 4507193 cached!
  ...
  invalidate_cache
  1.408584670163691 10540 4507193
  0.001336586195975542 10540 4507193 cached!
  0.0009196233004331589 10540 4507193 cached!
  0.0009176661260426044 10540 4507193 cached!
  ...
about 30s of low activity.

Without this patch, the UI is prone to freezing,
running behind, and I wouldn't be surprised if UI thread blocking
on Windows ends up crashing to some extent as the issue notes.
In the log, this manifests as a much slower but consistent
stream of full 1.3-1.4s updates during use,
and every time the window is focused.