cli: add max_cltv parameter to lnpay
Adds a `max_cltv` parameter to the `lnpay` cli command which allows to
specify the maximum total locktime of the payment. This is enabled by
constructing a custom `PaymentFeeBudget` object in the lnpay command and
passing it as argument to `LNWallet.pay_invoice()`.
Allowing to specify a `max_cltv` value can be useful for certain
usecases, e.g. see https://github.com/spesmilo/electrum/issues/10056.

Closes #10056