lnworker: make PaymentInfo dataclass, add more attributes
Adds the min_final_cltv_delta value to PaymentInfo so that it is
possible to verify it later when receiving the htlcs in lnpeer.
Currently the min_final_cltv_delta which was used in get_bolt11_invoice
for the lightning invoice was not persisted (except maybe through the
invoice in a Request object), so it was not possible to verify it when
receiving htlcs.
get_bolt11_invoice is now dependent on PaymentInfo being available, so
PaymentInfo has to be created first. This is done to prevent creating a
bolt11 invoice with different min_final_cltv_delta than the value stored
in PaymentInfo.

wallet_db: add upgrade for PaymentInfo dataclass

Version 60 to 61, converts the PaymentInfo tuple into the PaymentInfo
dataclass.

lnworker: make get_bolt11_invoice use PaymentInfo

This ensures that invoices can only be created if PaymentInfo has been
created first. It also ensures that the values in PaymentInfo are
similar to the values shared in the invoice.