qt: improve send_change_to_lightning feedback
Improves the GUI feedback given to the user when sending change to
lightning by showing different failure messages and the additional fee
charged for the swap.
Also considers the min/max swap amount announced
by the provider when deciding to add the swap dummy output in make tx
for improved reliability.