lnutil: add SETTLING state to RecvMPPResolution
this allows to have a clear distinction between a set that can still be
failed and a set that must be settled (not failed) and makes it possible to add more
safeguards. If we settle a htlc set only
partially due to some reason, e.g. crash, we shouldn't try to settle the
same htlcs on restart again, so they must be removed from the set.
However if the set would still be in COMPLETE state but with a partial
set of htlcs (some have been removed already before)
it is theoretically invalid and should be failed as sets in
COMPLETE state can still be failed. sets in SETTLED state should be
settled even if they wouldn't pass validation in their current state.