lnonion: store onion version and add getter methods
Adds some useful getter methods to `ProcessedOnionPacket` so it is not
neccessary to get everything by keys from the payloads dict.
Store the version of the `OnionPacket` so it can later be used to send
the correct error message if comparison fails instead of just raising
when a invalid version is passed to the constructor.