lnonion: add OnionParsingError, and util to OnionRoutingFailure
Adds OnionParsingError class which can be used for unparsable onions to
differentiate between a invalid onion and an onion which cannot even be
parsed as for unparsable onions there is a separate
`update_fail_malformed_htlc` rejection message.
Also adds `to_wire_message` util method to `OnionRoutingFailure` to get
the failure message for the wire from a `OnionRoutingFailure` as this
logic is only used for `OnionRoutingFailure` it seemed to make sense to
just add it as method.