change addresses for Alert and Cancellation
prefer change address for Alert Address and
Cancellation Address. When signing a transaction
that sends funds to a self-address on the main
derivation path (non-change), some hardware
wallets (i.e. Ledger) display this address as if
it belongs to an external wallet and as if the
amount sent there is will be deducted from the
wallet balance. This requires users to do an
extra step and verify that this destination
address is indeed theirs on the wallet,
and was not manipulated during the communication
with the hardware wallet.

Altough I don't agree with this approach, it
has some sense because typical transactions send
change funds only to addresses on the change
derivation path.

If we fail to find a change address, we fallback
to a main derivation-path address.